chore: move fast_thrift to util
