+++ pkg/utils/fast_thrift/fast_thrift.go
@@ -14,17 +14,19 @@
  * limitations under the License.
  */
 
-package thrift
+package fast_thrift
 
-// FastMarshal marshals the msg to buf. The msg should be generated by Kitex tool and implement thriftMsgFastCodec.
-func FastMarshal(msg thriftMsgFastCodec) []byte {
+import "github.com/cloudwego/kitex/pkg/remote/codec/thrift"
+
+// FastMarshal marshals the msg to buf. The msg should be generated by Kitex tool and implement ThriftMsgFastCodec.
+func FastMarshal(msg thrift.ThriftMsgFastCodec) []byte {
 	buf := make([]byte, msg.BLength())
 	msg.FastWriteNocopy(buf, nil)
 	return buf
 }
 
-// FastUnmarshal unmarshal the buf into msg. The msg should be generated by Kitex tool and implement thriftMsgFastCodec.
-func FastUnmarshal(buf []byte, msg thriftMsgFastCodec) error {
+// FastUnmarshal unmarshal the buf into msg. The msg should be generated by Kitex tool and implement ThriftMsgFastCodec.
+func FastUnmarshal(buf []byte, msg thrift.ThriftMsgFastCodec) error {
 	_, err := msg.FastRead(buf)
 	return err
 }
